requirements.txt
asgiref==3.3.1
attrs==21.4.0
bandit==1.7.0
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.2
distlib==0.3.4
dj-database-url==0.5.0
Django==3.1.7
django-filter==2.4.0
django-phonenumber-field==5.1.0
django-tables2==2.3.4
docopt==0.6.2
easygui==0.98.3
et-xmlfile==1.1.0
filelock==3.4.1
flake8==3.8.4
flake8-bugbear==21.3.2
gitdb==4.0.9
GitPython==3.1.20
google-api-core==2.7.1
google-api-python-client==2.46.0
google-auth==2.6.6
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.1
googleapis-common-protos==1.56.0
gunicorn==20.0.4
httplib2==0.20.4
idna==3.3
mccabe==0.6.1
mypy==0.931
mypy-extensions==0.4.3
oauth2client==4.1.3
oauthlib==3.2.0
openpyxl==3.0.7
patch==1.16
pbr==5.8.1
phonenumbers==8.12.21
pipreqs==0.4.11
platformdirs==2.4.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.21
PyDrive==1.3.1
PyDrive2==1.10.1
pyflakes==2.2.0
PyMsgBox==1.0.9
pyOpenSSL==22.0.0
pyparsing==3.0.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
smmap==5.0.0
sqlparse==0.4.1
stevedore==3.5.0
tk==0.1.0
tomli==1.2.3
typing_extensions==4.1.0
uritemplate==4.1.1
urllib3==1.26.9
virtualenv==20.14.1
whitenoise==5.2.0
yarg==0.1.9
